lints, even catched a bug
